Move send_request function from gatewat test utils to general test utils.
