+++ crates/papyrus_gateway/Cargo.toml
@@ -10,7 +10,6 @@ bincode = { version = "1.3.3" }
 papyrus_storage = { path = "../papyrus_storage" }
 futures-util = { version = "0.3.21" }
 jsonrpsee = { version = "0.13.1", features = ["full"] }
-reqwest = { version = "0.11", features = ["json"] }
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = { version = "1.0.81" }
 starknet_api = { git = "https://github.com/starkware-libs/starknet-api", rev = "9a1fca8" }
