+++ crates/test_utils/Cargo.toml
@@ -13,5 +13,6 @@ serde = { version = "1.0.130", features = ["derive"] }
 serde_json = { version = "1.0.81" }
 starknet_api = { git = "https://github.com/starkware-libs/starknet-api", rev = "9a1fca8" }
 tracing = { version = "0.1.37", features = ["log"] }
+reqwest = { version = "0.11", features = ["json"] }
 
 [dev-dependencies]
